✨ Feat: 리포트 카드 이미지 저장 UI 구현
