+++ packages/shared/design/shadows.css
@@ -12,5 +12,6 @@
   --shadow-current-location-button: 0 5.33px 13.33px 0 rgba(0, 0, 0, 0.1);
   --shadow-category-filter: -13px 0 2px 0 rgba(255, 255, 255, 0.65);
   --shadow-report-preference-card: 0 0 17px 0 rgba(0, 0, 0, 0.11);
+  --shadow-report-preference-save-card: 0 0 11px 0 rgba(0, 0, 0, 0.15);
   --shadow-chart-tooltip: 0 4px 12px 0 rgba(0, 0, 0, 0.18);
 }
